Refac: 확장 함수 위치 Lotto 클래스로 이동
